requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9,<3.0
pytesseract>=0.3.8
Pillow>=8.0.1
python-docx>=0.8.10
PyPDF2
llama-cpp-python